server/app/requirements.txt
asgiref==3.4.1
click==8.0.3
colorama==0.4.4
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
joblib==1.1.0
nltk==3.6.5
numpy==1.21.3
Pillow==8.4.0
psycopg2==2.9.1
PyJWT==2.3.0
pytz==2021.3
regex==2021.10.23
sqlparse==0.4.2
torch==1.10.0
torchvision==0.11.1
tqdm==4.62.3
typing-extensions==3.10.0.2
